+++ Libraries/LibGfx/Font.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2020, Andreas Kling <kling@serenityos.org>
+ * Copyright (c) 2020, Stephan Unverwerth <s.unverwerth@gmx.de>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -35,14 +35,9 @@
 
 namespace Gfx {
 
-enum FontTypes {
-    Default = 0,
-    LatinExtendedA = 1
-};
-
 // FIXME: Make a MutableGlyphBitmap buddy class for FontEditor instead?
 class GlyphBitmap {
-    friend class Font;
+    friend class BitmapFont;
 
 public:
     const unsigned* rows() const { return m_rows; }
@@ -75,108 +70,45 @@ private:
 
 class Font : public RefCounted<Font> {
 public:
-    NonnullRefPtr<Font> clone() const;
-    static NonnullRefPtr<Font> create(u8 glyph_height, u8 glyph_width, bool fixed, FontTypes type);
-
     static RefPtr<Font> load_from_file(const StringView& path);
-    bool write_to_file(const StringView& path);
 
-    ~Font();
+    virtual NonnullRefPtr<Font> clone() const = 0;
+    virtual ~Font() {};
 
-    u8 presentation_size() const { return m_presentation_size; }
-    void set_presentation_size(u8 size) { m_presentation_size = size; }
+    virtual u8 presentation_size() const = 0;
 
-    u16 weight() const { return m_weight; }
-    void set_weight(u16 weight) { m_weight = weight; }
+    virtual u16 weight() const = 0;
+    virtual GlyphBitmap glyph_bitmap(u32 code_point) const = 0;
 
-    GlyphBitmap glyph_bitmap(u32 code_point) const;
+    virtual u8 glyph_width(size_t ch) const = 0;
+    virtual int glyph_or_emoji_width(u32 code_point) const = 0;
+    virtual u8 glyph_height() const = 0;
+    virtual int x_height() const = 0;
 
-    u8 glyph_width(size_t ch) const { return m_fixed_width ? m_glyph_width : m_glyph_widths[ch]; }
-    int glyph_or_emoji_width(u32 code_point) const;
-    u8 glyph_height() const { return m_glyph_height; }
-    int x_height() const { return m_x_height; }
+    virtual u8 min_glyph_width() const = 0;
+    virtual u8 max_glyph_width() const = 0;
+    virtual u8 glyph_fixed_width() const = 0;
 
-    u8 min_glyph_width() const { return m_min_glyph_width; }
-    u8 max_glyph_width() const { return m_max_glyph_width; }
-    u8 glyph_fixed_width() const { return m_glyph_width; }
+    virtual u8 baseline() const = 0;
+    virtual u8 mean_line() const = 0;
 
-    u8 baseline() const { return m_baseline; }
-    void set_baseline(u8 baseline)
-    {
-        m_baseline = baseline;
-        update_x_height();
-    }
+    virtual int width(const StringView&) const = 0;
+    virtual int width(const Utf8View&) const = 0;
+    virtual int width(const Utf32View&) const = 0;
 
-    u8 mean_line() const { return m_mean_line; }
-    void set_mean_line(u8 mean_line)
-    {
-        m_mean_line = mean_line;
-        update_x_height();
-    }
+    virtual const String& name() const = 0;
 
-    int width(const StringView&) const;
-    int width(const Utf8View&) const;
-    int width(const Utf32View&) const;
+    virtual bool is_fixed_width() const = 0;
 
-    const String& name() const { return m_name; }
-    void set_name(String name) { m_name = move(name); }
+    virtual u8 glyph_spacing() const = 0;
 
-    bool is_fixed_width() const { return m_fixed_width; }
-    void set_fixed_width(bool b) { m_fixed_width = b; }
+    virtual int glyph_count() const = 0;
 
-    u8 glyph_spacing() const { return m_glyph_spacing; }
-    void set_glyph_spacing(u8 spacing) { m_glyph_spacing = spacing; }
+    virtual const String& family() const = 0;
 
-    void set_glyph_width(size_t ch, u8 width)
-    {
-        ASSERT(m_glyph_widths);
-        m_glyph_widths[ch] = width;
-    }
+    virtual String qualified_name() const = 0;
 
-    int glyph_count() const { return m_glyph_count; }
-
-    FontTypes type() { return m_type; }
-    void set_type(FontTypes type);
-
-    const String& family() const { return m_family; }
-    void set_family(String family) { m_family = move(family); }
-
-    String qualified_name() const;
-
-    const Font& bold_variant() const;
-
-private:
-    Font(String name, String family, unsigned* rows, u8* widths, bool is_fixed_width, u8 glyph_width, u8 glyph_height, u8 glyph_spacing, FontTypes type, u8 baseline, u8 mean_line, u8 presentation_size, u16 weight, bool owns_arrays = false);
-
-    static RefPtr<Font> load_from_memory(const u8*);
-    static size_t glyph_count_by_type(FontTypes type);
-
-    void update_x_height() { m_x_height = m_baseline - m_mean_line; };
-
-    String m_name;
-    String m_family;
-    FontTypes m_type;
-    size_t m_glyph_count { 256 };
-
-    unsigned* m_rows { nullptr };
-    u8* m_glyph_widths { nullptr };
-    MappedFile m_mapped_file;
-
-    u8 m_glyph_width { 0 };
-    u8 m_glyph_height { 0 };
-    u8 m_x_height { 0 };
-    u8 m_min_glyph_width { 0 };
-    u8 m_max_glyph_width { 0 };
-    u8 m_glyph_spacing { 0 };
-    u8 m_baseline { 0 };
-    u8 m_mean_line { 0 };
-    u8 m_presentation_size { 0 };
-    u16 m_weight { 0 };
-
-    bool m_fixed_width { false };
-    bool m_owns_arrays { false };
-
-    mutable RefPtr<Gfx::Font> m_bold_variant;
+    virtual const Font& bold_variant() const = 0;
 };
 
 }
